python-sidecar/requirements.txt
pillow
pytesseract
pyautogui
